replaced int with size_t

+++ PE#1/PE#1.cpp
@@ -7,8 +7,8 @@ using namespace std;
  */
 
 int main() {
-    int sum=0;
-    for(int i=0; i<1000; ++i){
+    size_t sum=0;
+    for(size_t i=0; i<1000; ++i){
         if (i%3==0 || i%5==0){
             sum = sum+i;
         }
